fix build error on ambiguous implicit convertion of long int on float/double in sqrt call (select double)

+++ src/plugin/base/constant_block_plugin.cpp
@@ -395,7 +395,7 @@ void MP_Constant_Block_Plugin_c::update_frame(unsigned long int frameIdx,
 			ip += (*pAmp);
 		}
 		if(!atEnd)
-			iter->corr[chanIdx] = ip/sqrt(filterLen);
+			iter->corr[chanIdx] = ip/sqrt((double)filterLen);
 		sum += ip * ip;
 	}
 	*maxCorr = sum/filterLen;
